+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Unreleased
 - only admin can update groups.
 - robocar number?
+- group scores?
+
 
 ## 0.4.0 - 2021-12-13
 - forbidden duplicate same user name in a group.